refactor: split downstream request execution
