+++ scripts/generate_quality_baseline.py
@@ -333,7 +333,7 @@ evidence for PR readiness, not a completion claim.
 | Tests | 77 Python test files at initial baseline; repo-native coverage gate existed | {len(_python_files(TESTS_DIR))} Python test files; {unit_tests_collected} tests collected in the latest baseline; OpenAPI gate logic has focused regression tests | Focused unit/integration coverage protects router, client, contract, middleware, service, and OpenAPI-governance refactors | Add more negative/security contract certification tests |
 | Security | Enterprise audit middleware, redaction tests, and upstream error mapping existed; abuse-control evidence was still a gap | Authorization checks, enterprise audit/redaction, and policy metadata are decomposed into dedicated modules; enterprise runtime and unmapped writes fail closed; correlation/trace input, downstream errors, response headers, and downstream base URLs are hardened with negative tests; threat-model and deployment policy evidence is pinned; Bandit and pip-audit remain green in baseline | Security behavior, deployment posture, unsafe-input handling, audit metadata handling, and abuse controls are easier to inspect and test without changing valid local-development semantics | Add gateway-backed token-validation evidence and final target-runtime configuration proof before release promotion |
 | Observability | HTTP, endpoint execution, supportability, freshness metrics, and correlation existed but needed consolidated docs | Observability docs, dashboard panels, alert definitions, runbook anchors, and endpoint/upstream metrics are covered by tests and baseline validation | Metrics/correlation posture is preserved through router and client decomposition, and operator response evidence is now governed | Keep alert thresholds aligned with production telemetry after deployment |
-| Resilience and performance | Downstream profiles declared timeout, connection, and keepalive limits, but operations created and closed a client per call | FastAPI lifespan owns reusable dependency-specific HTTP pools and closes them after entering draining posture; standalone/injected adapters remain supported; downstream timeout/pool/async polling env parsing is isolated in a focused helper; lotus-performance async returns payload validation and bounded failure construction are isolated from polling orchestration | Configured pooling now improves cross-request connection reuse and shutdown resource cleanup, and runtime profile/polling settings plus async failure behavior are easier to inspect and test | Add operation-specific retry only where idempotency and retry budgets are explicitly proven |
+| Resilience and performance | Downstream profiles declared timeout, connection, and keepalive limits, but operations created and closed a client per call | FastAPI lifespan owns reusable dependency-specific HTTP pools and closes them after entering draining posture; standalone/injected adapters remain supported; downstream timeout/pool/async polling env parsing is isolated in focused helpers; downstream request execution and upstream error observation are isolated from profile construction; lotus-performance async returns payload validation and bounded failure construction are isolated from polling orchestration | Configured pooling now improves cross-request connection reuse and shutdown resource cleanup, and runtime profile, request execution, polling settings, and async failure behavior are easier to inspect and test | Add operation-specific retry only where idempotency and retry budgets are explicitly proven |
 | Documentation and PR evidence | Baseline/reporting foundation was introduced with architecture, security, observability, runbook, wiki, and quality docs | `baseline_report.md`, `refactor_health_report.md`, `quality_scorecard.md`, and `final_pr_readiness.md` are updated with current measured movement and PR assembly evidence | Refactor progress is now auditable from generated reports and branch history | Final PR must attach current generated artifacts, CI status, and command evidence |
 
 ## Current Gate Snapshot
